7/24

+++ src/main/java/part1/Caluculator.java
@@ -4,8 +4,8 @@ public class Caluculator {
     public int multiply(int x, int y){
         return x * y;
     }
-    public int divide(int x, int y){
-        return x / y;
+    public float divide(int x, int y){
+        return (float) x / (float) y;
     }
 
 }
